backend/requirements.txt
# 后端依赖

# 基础依赖
Python>=3.9

# Web框架
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# 数据库
motor>=3.3.0  # 异步MongoDB驱动
pymongo>=4.4.0  # 同步MongoDB驱动

# 认证
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# 图像处理
Pillow>=10.0.0
imagehash>=4.3.0

# AI
google-generativeai>=0.3.0

# 环境变量
python-dotenv>=1.0.0

# 工具
tqdm>=4.65.0
requests>=2.31.0
python-dateutil>=2.8.2

# iCloud集成
pyicloud>=1.0.0